attached_assets/Pasted-Error-Element-type-is-invalid-expected-a-string-for-built-in-components-or-a-class-function-for-1747300047819.txt

Error: Element type is invalid: expected a string (for built-in components) or a class/function (for composite components) but got: undefined. You likely forgot to export your component from the file it's defined in, or you might have mixed up default and named imports.
    at renderElement (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:5835:15)
    at retryNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6092:31)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6042:11)
    at renderNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6499:18)
    at renderChildrenArray (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6375:9)
    at retryNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6116:13)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6042:11)
    at renderNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6499:18)
    at renderElement (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:5459:18)
    at retryNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6092:31)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6042:11)
    at renderNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6499:18)
    at renderChildrenArray (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6375:9)
    at retryNode (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6116:13)
    at renderNodeDestructive (file:///home/runner/workspace/node_modules/react-dom/cjs/react-dom-server.edge.development.js:6042:11)
